1. Fixed a few bugs in Makefile and reorganized constructive test directories 2. Wrote some simple unit tests on a few trivial methods of SimpleZoneAlertComputer'

+++ src/cpp/Plans/SimpleZoneAlertComputer.cpp
@@ -18,7 +18,7 @@ namespace zoneAlert {
 using std::shared_ptr;
 using std::make_shared;
 
-SimpleZoneAlertComputer::SimpleZoneAlertComputer(double lookahead) : ZoneAlertComputer(),
+SimpleZoneAlertComputer::SimpleZoneAlertComputer(int64_t lookahead) : ZoneAlertComputer(),
     lookaheadTime(lookahead),
     airVehicleConfigs(),
     boundaries(),
